requirements.txt
streamlit
pdfminer.six
scikit-learn
pandas
python-dotenv
requests
cohere